+++ app/src/scenes/account.js
@@ -8,12 +8,13 @@ import styled from "styled-components";
 import api from "../services/api";
 import { setYoung } from "../redux/auth/actions";
 import ErrorMessage, { requiredMessage } from "../scenes/inscription2023/components/ErrorMessageOld";
-import { getPasswordErrorMessage, translate, putLocation } from "../utils";
+import { getPasswordErrorMessage, translate } from "../utils";
 import validator from "validator";
 import AddressInputV2 from "../components/addressInputV2";
 import ModalConfirm from "../components/modals/ModalConfirm";
 import PasswordEye from "../components/PasswordEye";
 import { appURL } from "../config";
+import { putLocation } from "../services/api-adresse";
 
 export default function Account() {
   const young = useSelector((state) => state.Auth.young);
@@ -26,6 +27,7 @@ export default function Account() {
       if (!values.location || !values.location.lat || !values.location.lon) {
         values.location = await putLocation(values.city, values.zip);
       }
+      if (!values.location) return toastr.error("Il y a un soucis avec le nom de la ville ou/et le zip code");
       const { ok, code, data: young } = await api.put("/young", values);
       if (!ok) return toastr.error("Une erreur s'est produite :", translate(code));
       dispatch(setYoung(young));